requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.12.2
